requirements.txt
pyannote.audio
gradio